# dehaze-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "去雾系统后端 API 接口文档",
        "title": "Dehaze System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@dehaze.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/roles": {
            "post": {
                "description": "新增角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色接口"
                ],
                "summary": "新增角色",
                "parameters": [
                    {
                        "description": "角色表单对象",
                        "name": "roleForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bo.RoleFormBO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/roles/page": {
            "get": {
                "description": "获取角色分页列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色接口"
                ],
                "summary": "角色分页列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键字(角色名称/角色编码)",
                        "name": "keywords",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/common.PageResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{ids}": {
            "delete": {
                "description": "删除角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色接口"
                ],
                "summary": "删除角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "删除角色，多个以英文逗号(,)拼接",
                        "name": "ids",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{id}": {
            "put": {
                "description": "修改角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色接口"
                ],
                "summary": "修改角色",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "角色表单对象",
                        "name": "roleForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bo.RoleFormBO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{roleId}/form": {
            "get": {
                "description": "获取角色表单数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色接口"
                ],
                "summary": "角色表单数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.RoleFormBO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{roleId}/menuIds": {
            "get": {
                "description": "获取角色的菜单ID集合",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色接口"
                ],
                "summary": "获取角色的菜单ID集合",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer",
                                                "format": "int64"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{roleId}/menus": {
            "put": {
                "description": "分配菜单(包括按钮权限)给角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色接口"
                ],
                "summary": "分配菜单(包括按钮权限)给角色",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "菜单ID列表",
                        "name": "menuIds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{roleId}/status": {
            "put": {
                "description": "修改角色状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色接口"
                ],
                "summary": "修改角色状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "状态(1:启用;0:禁用)",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bo.RoleFormBO": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "description": "角色编码",
                    "type": "string"
                },
                "dataScope": {
                    "description": "数据权限",
                    "type": "integer"
                },
                "id": {
                    "description": "角色ID",
                    "type": "integer"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "角色状态(1-正常；0-停用)",
                    "type": "integer"
                }
            }
        },
...
        "common.PageResult": {
            "type": "object",
            "properties": {
                "list": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}